Live Refresh Markdown Preview
I used a the `codemirror` package as the markdown editor with stimulus
to trigger form updates with the change sets. The change sets are
applied to a copy of the data stored in the `WorkingDocument` record.
Then the markdown is rendered to HTML and Turbo updates the preview card
on the clientside.

By using `find_or_initialize_by`, we avoid creating a new
`WorkingDocument` record until the user actually makes an edit to the
document. This is important because we don't want to create a record for
every issue that is opened in the repo, only the ones that are actually
being edited.

We can explore other options for cost saving in the future, ideally we
don't need to persist the document body in the database at all.

* The minimum viable solution is to use a background job to delete old
  `WorkingDocument` records.

* The ideal solution is to not need to persist the document body at all.
  * We can send the document body to the server to render the preview
    card.
  * We can use `ActionCable` to broadcast the `ChangeSet` to all
    connected clients so they can update their codemirror editors.
  * We could also use `ActionCable` to broadcast the updated preview
    card.
  * :warn: Updates could be lost on page refresh, need to investigate
    local storage options.
  * Newly connected clients would need to fetch the latest document body
    state via `ActionCable` from another connected client.
  * When multiple client rejoin with different copies of the body we
    need to generate and resolve/surface merge conflicts.

Using the ChangeSet to update comment positions is the trickiest part of
this application so I'm going to focus on that first. This comment
should be sufficient documentation for next steps in this problem
domain.