Use helper for github color based on issue status
The `content` array in `tailwind.config.js` lists the files that
tailwind will scan to find used styles. It already includes the
`app/helpers` directory since those are made avaliable when rendering a
view template.

Previously, I attempted to add this logic to the Issue model directly
with an `Issue#color` method. However, tailwind would strip the color
from the bundle because it thought it wasn't used anywhere in the view.
Rather than add the `app/models` to the tailwind config I've used the
view helper.